A65
The program has a swap function that only swaps two values by using a temp variable. Then there is an overloading function that swaps three values. It does this by using the swaptwo function twice in order to switch the first and the second and the first and third one after each other. These two functions are void because the function changes the reference of those variables.